components/job_creation/python/requirements.txt
pandas==0.25.1
geopandas==0.5.1
Shapely==1.6.4.post2
PyYAML==5.2
numpy==1.16.2
